1.0.5

+++ modules/Accounting/Http/Controllers/AccountingSettingCategoryController.php
@@ -3,21 +3,22 @@
 namespace Modules\Accounting\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Illuminate\Contracts\Support\Renderable;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Routing\Controller;
-use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Contracts\Support\Renderable;
 use Modules\Accounting\Models\AccountingEntryCategory;
+use Illuminate\Foundation\Validation\ValidatesRequests;
 
 /**
- * @class AccountingEntryCategoryController
+ * @class AccountingSettingCategoryController
  * @brief Controlador para la gestion las categorias de origen
  *
  * Clase que gestiona las categorias de origen
  *
- * @author Juan Rosas <jrosas@cenditel.gob.ve | juan.rosasr01@gmail.com>
- * @license<a href='http://conocimientolibre.cenditel.gob.ve/licencia-de-software-v-1-3/'>
- *              LICENCIA DE SOFTWARE CENDITEL
- *          </a>
+ * @author Juan Rosas <jrosas@cenditel.gob.ve> | <juan.rosasr01@gmail.com>
+ *
+ * @license
+ *     [LICENCIA DE SOFTWARE CENDITEL](http://conocimientolibre.cenditel.gob.ve/licencia-de-software-v-1-3/)
  */
 class AccountingSettingCategoryController extends Controller
 {
@@ -26,11 +27,13 @@ class AccountingSettingCategoryController extends Controller
     /**
      * Define la configuración de la clase
      *
-     * @author Juan Rosas <jrosas@cenditel.gob.ve | juan.rosasr01@gmail.com>
+     * @author Juan Rosas <jrosas@cenditel.gob.ve> | <juan.rosasr01@gmail.com>
+     *
+     * @return void
      */
     public function __construct()
     {
-        /** Establece permisos de acceso para cada método del controlador */
+        // Establece permisos de acceso para cada método del controlador
         $this->middleware('permission:accounting.setting.category.store', ['only' => ['store']]);
         $this->middleware('permission:accounting.setting.category.update', ['only' => ['update']]);
         $this->middleware('permission:accounting.setting.category.delete', ['only' => 'destroy']);
@@ -44,9 +47,11 @@ class AccountingSettingCategoryController extends Controller
     /**
      * Crea una nueva categorias de origen de asiento contable
      *
-     * @author Juan Rosas <jrosas@cenditel.gob.ve | juan.rosasr01@gmail.com>
-     * @param  Request $request Objeto con datos de la petición realizada
-     * @return Renderable
+     * @author Juan Rosas <jrosas@cenditel.gob.ve> | <juan.rosasr01@gmail.com>
+     *
+     * @param  Request $request Datos de la petición realizada
+     *
+     * @return JsonResponse
      */
     public function store(Request $request)
     {
@@ -65,10 +70,12 @@ class AccountingSettingCategoryController extends Controller
     /**
      * Actualiza los datos de la categoria de origen
      *
-     * @author Juan Rosas <jrosas@cenditel.gob.ve | juan.rosasr01@gmail.com>
-     * @param  Request $request Objeto con datos de la petición realizada
+     * @author Juan Rosas <jrosas@cenditel.gob.ve> | <juan.rosasr01@gmail.com>
+     *
+     * @param  Request $request Datos de la petición realizada
      * @param  integer $id      Identificador de la categoria de origen a modificar
-     * @return Renderable
+     *
+     * @return JsonResponse
      */
     public function update(Request $request, $id)
     {
@@ -76,10 +83,7 @@ class AccountingSettingCategoryController extends Controller
             'name'    => ['required', 'string'],
             'acronym' => ['required', 'string'],
         ]);
-        /**
-         * [$record registro de conversión a editar]
-         * @var AccountingEntryCategory
-         */
+        /* registro de conversión a editar */
         $record          = AccountingEntryCategory::find($id);
         $record->name    = $request['name'];
         $record->acronym = $request['acronym'];
@@ -93,22 +97,28 @@ class AccountingSettingCategoryController extends Controller
     /**
      * Elimina una categoria de origen de asiento contable
      *
-     * @author Juan Rosas <jrosas@cenditel.gob.ve | juan.rosasr01@gmail.com>
+     * @author Juan Rosas <jrosas@cenditel.gob.ve> | <juan.rosasr01@gmail.com>
+     *
      * @param  integer $id Identificador de la categoria de origen a eliminar
-     * @return Renderable
+     *
+     * @return JsonResponse
      */
     public function destroy($id)
     {
         $category = AccountingEntryCategory::with('accountingEntries')->find($id);
 
         if ($category) {
-            /**
-             * validar si no esta relacionada con algun asiento es permitido eliminarla
-             */
+            /* validar si no esta relacionada con algun asiento es permitido eliminarla */
             if (!is_null($category->accountingEntries) && count($category->accountingEntries) > 0) {
                 return response()->json([
                     'error'   => true,
                     'message' => 'El registro no se puede eliminar, debido a que existen asientos contables que la usan.'
+                ], 200);
+            }
+            if ($category->acronym == 'ARA') {
+                return response()->json([
+                    'error'   => true,
+                    'message' => 'El registro no se puede eliminar, debido a que es una categoría predefinida.'
                 ], 200);
             }
             $category->delete();
@@ -117,10 +127,11 @@ class AccountingSettingCategoryController extends Controller
     }
 
     /**
-     * Obtiene los datos de las entidades bancarias
+     * Obtiene los datos de las categorías
      *
      * @author Ing. Roldan Vargas <rvargas@cenditel.gob.ve> | <roldandvg@gmail.com>
-     * @return \Illuminate\Http\JsonResponse Devuelve un JSON con listado de las entidades bancarias
+     *
+     * @return JsonResponse Devuelve un JSON con listado de las categorías
      */
     public function getCategories()
     {
